feat: "엔티티 설계 및 연관관계 수정, 테스트 파일 추가"

+++ backend/src/main/java/a609/backend/controller/TripController.java
@@ -25,7 +25,7 @@ public class TripController {
     }
 
     @GetMapping("/showTripInfo/{tripId}")
-    public ResponseEntity<Map<String,Object>> showTripInfo(@PathVariable int tripId){
+    public ResponseEntity<Map<String,Object>> showTripInfo(@PathVariable Long tripId){
         Map<String, Object> resultMap = new HashMap<>();
         Trip tripInfo = tripService.showTripInfo(tripId);
         resultMap.put("tripInfo", tripInfo);
@@ -34,7 +34,7 @@ public class TripController {
     }
 
     @PutMapping("/update/{tripId}")
-    public ResponseEntity<Map<String,Object>> updateTrip(@PathVariable int tripId,@RequestBody Trip trip){
+    public ResponseEntity<Map<String,Object>> updateTrip(@PathVariable Long tripId,@RequestBody Trip trip){
         Map<String, Object> resultMap = new HashMap<>();
         tripService.updateTrip(tripId,trip);
         resultMap.put("message", "Success");
@@ -43,7 +43,7 @@ public class TripController {
 
     //여행 전체 삭제
     @DeleteMapping("/trip/{tripId}")
-    public ResponseEntity<Map<String,Object>> deleteTrip(@PathVariable int tripId){
+    public ResponseEntity<Map<String,Object>> deleteTrip(@PathVariable Long tripId){
         Map<String, Object> resultMap = new HashMap<>();
         tripService.deleteTrip(tripId);
