+++ backend/src/main/java/a609/backend/service/TripService.java
@@ -7,13 +7,14 @@ import java.util.List;
 
 public interface TripService {
 
-    Trip showTripInfo(int tripId);
+    Trip showTripInfo(Long tripId);
 //    List<Trip> showTripList(int userId);
 
     void registerTrip(Trip trip,String userId);
-    void updateTrip(int tripId,Trip trip);
-    void deleteTrip(int tripId);
+    void updateTrip(Long tripId,Trip trip);
+    void deleteTrip(Long tripId);
 
+    //??이거 무슨 메소드더라???????
     void deleteUserTrip(int userId);
 
 }
